BE/STS_BE/requirement.txt
Django==6.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==23.5
openai==1.12.0
python-dotenv==1.0.1
